Add all buttons in separate file
