+++ src/components/Calculator.js
@@ -1,9 +1,9 @@
 import { React, Component } from 'react';
-import CalculatorOperations from './CalculatorOperations';
-import Numbers from './Numbers';
-import SimpleOperations from './SimpleOperations';
+import CalculatorOperations from './Buttons/CalculatorOperations';
+import Numbers from './Buttons/Numbers';
+import SimpleOperations from './Buttons/SimpleOperations';
 import classes from './Calculator.module.css';
-import Input from './Input';
+import Input from './Buttons/Input';
 // eslint-disable-next-line react/prefer-stateless-function
 class Calculator extends Component {
   render() {
